requirements.txt
# LLM and AI agent framework
langchain==0.1.12
langchain-groq
groq
llama-index

# backend(API)
fastapi==0.109.2
uvicorn==0.27.0.post1

# Database
sqlalchemy==2.0.27
sqlite-utils==3.35

# Environment Variables
python-dotenv==1.0.1

# Logging and Debugging
loguru==0.7.2

# Task Scheduling
apscheduler==3.10.4

# CLI Interaction
typer==0.9.0

# Streamlit
streamlit==1.31.1